content/canvas/test/test_2d.imageData.get.order.alpha.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.imageData.get.order.alpha</title>
<!-- Testing: getImageData() returns A in the fourth component -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ctx.fillStyle = 'rgba(0, 0, 0, 0.5)';
ctx.fillRect(0, 0, 100, 50);
var imgdata = ctx.getImageData(0, 0, 10, 10);
ok(imgdata.data[3] < 200, "imgdata.data[\""+(3)+"\"] < 200");
ok(imgdata.data[3] > 100, "imgdata.data[\""+(3)+"\"] > 100");

SimpleTest.finish();

});
</script>
